+++ src/osu/data/gosu/gosumemorydatawrapper.h
